requirements.txt
httpx>=0.23.0
pyyaml>=6.0
scipy>=1.9.0
python-dotenv>=1.0.0
